Factor functions out of ComputeTimeDerivative
Moves some functions and metafunctions into separate files to keep the
ComputeTimeDerivative file from growing even more.
